Backend/services/user_service/requirements.txt
fastapi==0.110.0
uvicorn==0.29.0
sqlalchemy==2.0.30
pydantic==2.7.1
passlib[bcrypt]==1.7.4
python-jose==3.3.0
alembic==1.13.1
requests==2.31.0
itsdangerous==2.1.2
authlib==1.3.1
httpx==0.27.0
asyncpg==0.29.0
